New executable "ester evol" to calculate a pseudo-evolution. At this moment this evolution is calculated just varying Xc while maintaining constant M and Lz. This behaviour is provisional and will change in future versions. To use it we must specify an input model (using -i or -input_file) calculated using ester 1d or ester 2d:          ester evol -i input_model Other valid parameters are:    -dXc : decrement in Xc (default: 0.05)    -Xcmin : minimum value of Xc (default: 0.05)    -o outfile : The default is outfile=input_file and other parameters inherited from ester 2d (-output_mode, -verbose, ...). However, it is not possible to change model parameters with ester evol (as mass, resolution, etc.). The output is a series of files output_XXXX, where XXXX is an index number, containing the models at different stages.
Included functions to calculate Lz, Lzcore and Mcore. This values are also accesible using ester output and within the python module.

In the python module, I have included a function star_evol to read the series of models created with ester evol. Use it as:
   A=star_evol(file)
Then, it reads the files file_XXXX and returns a list of star2d objects. It is possible to get the variation of a single parameter during the evolution doing, for example:
   Rp=[a.Rp for a in A]


git-svn-id: http://ester-project.googlecode.com/svn/trunk@101 dc0d849c-13d9-9696-9681-8a32c59e610f